refactor: try to fix many errors
