+++ .github/CODEOWNERS
@@ -22,4 +22,3 @@
 
 # Tyler & Kyra
 /services @Tylertron1998 @kyranet
-
