+++ Dockerfile
@@ -49,7 +49,6 @@ FROM base AS runner
 ENV NODE_ENV="production"
 ENV NODE_OPTIONS="--enable-source-maps --max_old_space_size=4096"
 
-COPY --chown=node:node scripts/workerTsLoader.js scripts/workerTsLoader.js
 COPY --chown=node:node src/.env src/.env
 COPY --chown=node:node --from=builder /usr/src/app/dist dist
