+++ package.json
@@ -98,11 +98,6 @@
 	},
 	"devDependencies": {
 		"0x": "^5.2.1",
-		"@babel/core": "^7.18.2",
-		"@babel/plugin-proposal-class-properties": "^7.17.12",
-		"@babel/plugin-proposal-decorators": "^7.18.2",
-		"@babel/preset-env": "^7.18.2",
-		"@babel/preset-typescript": "^7.17.12",
 		"@commitlint/cli": "^17.0.2",
 		"@commitlint/config-conventional": "^17.0.2",
 		"@sapphire/eslint-config": "^4.3.6",
@@ -124,12 +119,10 @@
 		"eslint-config-prettier": "^8.5.0",
 		"eslint-plugin-prettier": "^4.0.0",
 		"husky": "^8.0.1",
-		"jest-mock-random": "^1.1.1",
 		"lint-staged": "^13.0.0",
 		"nock": "^13.2.6",
 		"prettier": "^2.6.2",
 		"pretty-quick": "^3.1.3",
-		"ts-node": "^10.8.1",
 		"tsc-watch": "^5.0.3",
 		"typescript": "^4.7.3",
 		"vitest": "^0.14.2"
@@ -160,5 +153,8 @@
 		}
 	},
 	"prettier": "@sapphire/prettier-config",
-	"packageManager": "yarn@3.2.1"
+	"packageManager": "yarn@3.2.1",
+	"volta": {
+		"node": "16.15.1"
+	}
 }
