+++ tests/common/MockInstances.ts
@@ -2,7 +2,16 @@ import { LanguageKeys } from '#lib/i18n/languageKeys';
 import { SkyraCommand } from '#lib/structures';
 import { CLIENT_OPTIONS } from '#root/config';
 import { SapphireClient } from '@sapphire/framework';
-import { APIChannel, APIGuild, APIGuildMember, APIRole, APIUser, ChannelType, GuildFeature, GuildNSFWLevel } from 'discord-api-types/v9';
+import {
+	ChannelType,
+	GuildFeature,
+	GuildNSFWLevel,
+	type APIChannel,
+	type APIGuild,
+	type APIGuildMember,
+	type APIRole,
+	type APIUser
+} from 'discord-api-types/v9';
 import { Guild, GuildMember, Role, TextChannel, User } from 'discord.js';
 import { resolve } from 'node:path';
 
@@ -16,6 +25,7 @@ export const userData: APIUser = {
 };
 
 export function createUser(data: Partial<APIUser> = {}) {
+	// @ts-expect-error Private constructor:
 	return new User(client, { ...userData, ...data });
 }
 
@@ -30,6 +40,7 @@ export const guildMemberData: APIGuildMember = {
 };
 
 export function createGuildMember(data: Partial<APIGuildMember> = {}, g: Guild = guild) {
+	// @ts-expect-error Private constructor:
 	return new GuildMember(client, { ...guildMemberData, ...data, user: { ...guildMemberData.user, ...data.user! } }, g);
 }
 
@@ -45,6 +56,7 @@ export const roleData: APIRole = {
 };
 
 export function createRole(data: Partial<APIRole> = {}, g: Guild = guild) {
+	// @ts-expect-error Private constructor:
 	const role = new Role(client, { ...roleData, ...data }, g);
 	g.roles.cache.set(role.id, role);
 	return role;
@@ -71,6 +83,7 @@ export const guildData: APIGuild = {
 		GuildFeature.InviteSplash,
 		GuildFeature.Community
 	],
+	hub_type: null,
 	max_members: 100000,
 	max_presences: null,
 	max_video_channel_users: 25,
@@ -78,6 +91,7 @@ export const guildData: APIGuild = {
 	nsfw_level: GuildNSFWLevel.Default,
 	owner_id: '242043489611808769',
 	preferred_locale: 'en-US',
+	premium_progress_bar_enabled: false,
 	premium_subscription_count: 3,
 	premium_tier: 1,
 	public_updates_channel_id: '700806874294911067',
@@ -95,6 +109,7 @@ export const guildData: APIGuild = {
 };
 
 export function createGuild(data: Partial<APIGuild> = {}) {
+	// @ts-expect-error Private constructor:
 	const g = new Guild(client, { ...guildData, ...data });
 	client.guilds.cache.set(g.id, g);
 	return g;
@@ -117,6 +132,7 @@ export const textChannelData: APIChannel = {
 };
 
 export function createTextChannel(data: Partial<APIChannel> = {}, g: Guild = guild) {
+	// @ts-expect-error Private constructor:
 	const c = new TextChannel(guild, { ...textChannelData, ...data });
 	g.channels.cache.set(c.id, c);
 	g.client.channels.cache.set(c.id, c);
@@ -142,7 +158,7 @@ addCommand(
 		},
 		{
 			description: LanguageKeys.Commands.General.PingDescription,
-			extendedHelp: LanguageKeys.Commands.General.PingExtended,
+			detailedDescription: LanguageKeys.Commands.General.PingExtended,
 			aliases: ['pong'],
 			fullCategory: ['General']
 		}
@@ -159,7 +175,7 @@ addCommand(
 		},
 		{
 			description: LanguageKeys.Commands.Social.BalanceDescription,
-			extendedHelp: LanguageKeys.Commands.Social.BalanceExtended,
+			detailedDescription: LanguageKeys.Commands.Social.BalanceExtended,
 			aliases: ['bal'],
 			fullCategory: ['Currency']
 		}
@@ -176,7 +192,7 @@ addCommand(
 		},
 		{
 			description: LanguageKeys.Commands.Tools.DefineDescription,
-			extendedHelp: LanguageKeys.Commands.Tools.DefineExtended,
+			detailedDescription: LanguageKeys.Commands.Tools.DefineExtended,
 			aliases: ['def', 'definition', 'defination', 'dictionary'],
 			fullCategory: ['Tools', 'Dictionary']
 		}
