+++ tests/common/image.png
