+++ tests/lib/database/settings/structures/PermissionNodeManager.test.ts
@@ -1,6 +1,6 @@
+import { createGuild, createGuildMember, createRole, createUser, roleData } from '#common/MockInstances';
 import { GuildEntity, PermissionNodeAction, PermissionsNode } from '#lib/database';
 import { LanguageKeys } from '#lib/i18n/languageKeys';
-import { createGuild, createGuildMember, createRole, createUser, roleData } from '#mocks/MockInstances';
 import { UserError } from '@sapphire/framework';
 import type { Guild } from 'discord.js';
