+++ tests/lib/util/functions/emojis/getEmojiReactionFormat.test.ts
@@ -1,4 +1,4 @@
-import { encodedBunnyTwemoji, serializedAnimatedSkyraGlasses, serializedStaticSkyra } from '#mocks/constants';
+import { encodedBunnyTwemoji, serializedAnimatedSkyraGlasses, serializedStaticSkyra } from '#common/constants';
 import { getEmojiReactionFormat } from '#utils/functions';
 
 describe('getEmojiReactionFormat', () => {
