+++ tests/lib/util/functions/emojis/getEmojiString.test.ts
@@ -1,4 +1,4 @@
-import { bunnyTwemoji, encodedBunnyTwemoji, serializedAnimatedSkyraGlasses, serializedStaticSkyra } from '#mocks/constants';
+import { bunnyTwemoji, encodedBunnyTwemoji, serializedAnimatedSkyraGlasses, serializedStaticSkyra } from '#common/constants';
 import { getEmojiString } from '#utils/functions';
 
 describe('getEmojiString', () => {
