+++ tests/tsconfig.json
@@ -4,11 +4,12 @@
 		"noEmit": true,
 		"incremental": false,
 		"baseUrl": ".",
+		"types": ["vitest/globals"],
 		"paths": {
 			"#utils/*": ["../src/lib/util/*"],
 			"#lib/*": ["../src/lib/*"],
 			"#languages": ["../src/languages/index"],
-			"#mocks/*": ["./mocks/*"],
+			"#common/*": ["./common/*"],
 			"#root/*": ["../src/*"]
 		},
 		"types": ["vitest/globals"]
